on PC without warnings

+++ include/daScript/simulate/aot_builtin_string.h
@@ -36,7 +36,7 @@ namespace das {
     __forceinline char * format ( const char * fmt, TT value, Context * context ) {
         char buf[256];
         snprintf(buf, 256, fmt, value);
-        return context->heap.allocateString(buf, strlen(buf));
+        return context->heap.allocateString(buf, uint32_t(strlen(buf)));
     }
 
     template <typename TT>
